Fix missing underscore dependency
File `quorum-overloads` from `interface-adapter` imported the underscore library, even though it was not declared on the package.json as a dependency. This caused a `require` of the package (or any other package that depended on it) to fail.

Instead of installing underscore, since other truffle packages already use lodash, this commit changes underscore to lodash, and declares it as a dependency on the package.json.