Makes `config` available in the migration context.
This will allow a migration to be able to interact with any of the Truffle packages using the same configuration that the migration is using to run.  One example use case (mine) is to compile dynamic contracts based on the result of migration (e.g., hard-code an address in a contract as a gas optimization rather than passing it as a constructor).

Along the way I also cleaned up options vs config a little bit, once `config` is resolved it should be used rather than `options` since it may have different property values once full config resolution is complete.