Use license from config/package for new contracts
Even when you have a license listed in your project configuration it wasn't
being used for new contracts, with this small change the license listed in
config will be used.

How to see it working?

1. `truffle init someproject`
2. `cd someproject`
3. Edit `truffle-config.json` and add a key `license: "GPL"` or whatever
   license you want to use.
4. `truffle create contract Contract`
5. `head -n1 contracts/Contract.sol` should show the new license.

In case the license is not present in the `truffle-config.json` file it will
check for it in the `package.json` file. The process to check this is the same
than the previously mentioned but writting the license in `package.json`.