core: Change repl startup logic
This pr changes 1) how the repl context is instantiated and 2) removes annoying
`undefined` from repl output

This change is necessary to capture the globals that the repl includes in the
context. Without it there is no console.log and other useful globals needed for
a good user experince.

The motivation for th previous change was to allow the truffle context
variables to be hydrated before processing user input which is needed for our
tests. We can still achieve this by coordinating when to send the input from
tests. The sequence of events is as follows:

child Process:
  Repl starts with nodes context and prompt = ''
  Repl loads truffle specific variables into context
  Repl sets the prompt to '(develop)>'

test runner:
  Scans child's sdtout looking for '(develop>' before sending input