Allow the display of the number of instruction to be configurable
This PR addresses #4437 by making the display of the number of
instructions and source file configurable.

Use `p -<num>|+<num>` to configure number of instructions before and
after the current instruction. The default is to display 3 instructions
before and 3 instructions after the current instruction.

Use `l -<num>|+<num>` to configure the number of line of source file
before and after the current line of source file. The default is to
display 5 lines before and 3 lines after the current line.