Standardize @types/node for v12 LTS
This commit standardizes all occurrences of the `@types/node` to version
range `~12.12.0`. This version range gives us the latest patch release
of the node types that apply to the earliest LTS release of node v12
(v12.13.0).

This change uncovered two node version compatibility issues that had
been undetected prior to this change:

1. The version of `axios` that we use (v0.27.2) depends on the
   `AbortController` object that was introduced in node v15.0.0.
2. The version of the `conf` package that was used by @truffle/config
   was required a bugfix to build properly with earlier versions of
   `@node/types`.

As a result, this change includes the following fixes:

1. Includes the `node-abort-controller` polyfill and adds a bit of logic
   to conditionally install it on node versions where the
   `AbortController` is not yet present.
2. Bumps the version of the `conf` package to `10.1.2`, which includes a
   types fix that makes the package compatible with the earliest LTS v12
   release.